app/requirements.txt
PyMySQL==1.1.2
python-dotenv==1.2.2
SQLAlchemy==2.0.48
streamlit==1.55.0